refactor: abstracted swizzled components to common config
