+++ packages/common-config/src/components/theme/index.ts
@@ -1,0 +1,4 @@
+// Theme Components
+export { default as Root } from "./Root";
+export { default as DocRootLayoutSidebarExpandButton } from "./DocRoot/Layout/Sidebar/ExpandButton";
+export { default as DocSidebarDesktopCollapseButton } from "./DocSidebar/Desktop/CollapseButton";
