+++ packages/common-config/static/common/styles/global.css
@@ -805,35 +805,6 @@ a:hover {
   vertical-align: middle;
 }
 
-/* Font for category titles */
-.theme-doc-sidebar-item-category
-  .menu__list-item-collapsible
-  .menu__link[href^="/docs/processes"],
-.theme-doc-sidebar-item-category
-  .menu__list-item-collapsible
-  .menu__link[href^="/docs/development"],
-.theme-doc-sidebar-item-category
-  .menu__list-item-collapsible
-  .menu__link[href^="/docs/security"],
-.theme-doc-sidebar-item-category
-  .menu__list-item-collapsible
-  .menu__link[href^="/docs/testing"] {
-  font-family:
-    "Inter",
-    system-ui,
-    -apple-system,
-    sans-serif !important;
-  font-weight: 700 !important;
-  line-height: 1 !important;
-  letter-spacing: 0.03em;
-  color: var(--wonderland-accent);
-  text-transform: uppercase;
-  font-size: 0.8rem !important;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
 .menu__list-item-collapsible .menu__link:hover::before {
   opacity: 1;
 }
