+++ sites/template/src/theme/DocSidebar/Desktop/CollapseButton/index.tsx
@@ -1,0 +1,1 @@
+export { DocSidebarDesktopCollapseButton as default } from '@handbook/common-config/theme';
