+++ sites/wonderland/src/theme/DocRoot/Layout/Sidebar/ExpandButton/index.tsx
@@ -1,55 +1,1 @@
-import React, { type ReactNode } from "react";
-import { translate } from "@docusaurus/Translate";
-import type { Props } from "@theme/DocRoot/Layout/Sidebar/ExpandButton";
-
-import styles from "./styles.module.css";
-
-// Custom Sidebar Icon Component
-function SidebarIcon({ className }: { className?: string }) {
-  return (
-    <svg
-      width="22"
-      height="18"
-      viewBox="0 0 22 18"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={className}
-    >
-      <path
-        d="M15 13L15 14M15 16.5L15 2L15 0.999999M15 4L15 5M15 0.999999L5.8 0.999999C4.11984 0.999999 3.27977 0.999999 2.63803 1.32698C2.07354 1.6146 1.6146 2.07354 1.32698 2.63803C1 3.27976 1 4.11984 1 5.8L1 12.2C1 13.8802 1 14.7202 1.32698 15.362C1.6146 15.9265 2.07354 16.3854 2.63803 16.673C3.27976 17 4.11984 17 5.8 17L16.2 17C17.8802 17 18.7202 17 19.362 16.673C19.9265 16.3854 20.3854 15.9265 20.673 15.362C21 14.7202 21 13.8802 21 12.2L21 5.8C21 4.11984 21 3.27976 20.673 2.63803C20.3854 2.07354 19.9265 1.6146 19.362 1.32698C18.7202 1 17.8802 1 16.2 1L15 0.999999ZM7.5 12L10.5 9L7.5 6"
-        stroke="currentColor"
-        strokeWidth="1.5"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-}
-
-export default function DocRootLayoutSidebarExpandButton({
-  toggleSidebar,
-}: Props): ReactNode {
-  return (
-    <div
-      className={styles.expandButton}
-      title={translate({
-        id: "theme.docs.sidebar.expandButtonTitle",
-        message: "Expand sidebar",
-        description:
-          "The ARIA label and title attribute for expand button of doc sidebar",
-      })}
-      aria-label={translate({
-        id: "theme.docs.sidebar.expandButtonAriaLabel",
-        message: "Expand sidebar",
-        description:
-          "The ARIA label and title attribute for expand button of doc sidebar",
-      })}
-      tabIndex={0}
-      role="button"
-      onKeyDown={toggleSidebar}
-      onClick={toggleSidebar}
-    >
-      <SidebarIcon className={styles.expandButtonIcon} />
-    </div>
-  );
-}
+export { DocRootLayoutSidebarExpandButton as default } from "@handbook/common-config/theme";
