+++ sites/wonderland/src/theme/Root/index.tsx
@@ -1,5 +1,5 @@
-import React from "react";
+import { Root as SharedRoot } from "@handbook/common-config/theme";
 
 export default function Root({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
+  return <SharedRoot showDisclaimer={false}>{children}</SharedRoot>;
 }
